Restructure rdm launch polling loop to be based on elapsed time
The loop was previously based on the sum of presumed sleep delay times,
which could be much greater than the actual amount of elapsed time.

Also, add a timeout to rc --is-indexing calls in order to prevent too many
requests from being queued up if rdm is slow to respond.