Reload project if compile_commands.json was deleted and later recreated
RTags deletes the entire index for a project when compile_commands.json
is deleted. If compile_commands.json is later recreated, RTags does not
automatically reload it.

When the workspace folder for a project is opened, and
compile_commands.json is present, but the index is absent from RTags,
then reload compile_commands.json.