Apply symbol kind filters to symbol highlighting
In some cases, rc --all-references returns references to symbols in
different categories from the queried symbol. For example, rc
--all-references returns macro expansions when querying symbols that are
present in the macro, and it returns variable declarations that are also
constructor calls when querying constructors.

This change ensures that only symbols in the appropriate categories are
highlighted. The same effect can't be achieved by using --rename,
because that option can also return reference to symbols in different
categories.