Cache configuration settings locally
Compare new config against cached config to determine when to reload the
window, instead of relying on ConfigurationChangeEvent.affectsConfiguration.
The latter returns true when a folder has been opened or closed, but the
existing settings have not changed.

As a side effect, do not reload the window when resource-scoped
settings have changed, except for rtags.misc.compilationDatabaseDirectory.