Partially restore workaround from 1806d87573ad53d18ee96d0d7d200a157a31f0ec
The rtags.misc.compilationDatabaseDirectory setting changes when the
associated folder is closed. The folder must be removed from
workspaceFolders before proceeding so that the setting is not even
looked at for this folder. There's no way of detecting whether the
setting changed due to the folder being closed, or user input.