Take action when the value of misc.compilationDatabaseRecursiveSearch changes
Store full CompileCommandsInfo objects in the workspace storage so that
the previous value of misc.compilationDatabaseRecursiveSearch is
available when loading compilation databases. Prompt the user to delete
recursive compilation databases when the value of this setting changes
from true to false.

Also, rename the workspace storage from rtags.dirtyProjectPaths to
rtags.dirtyWorspaceInfo to better represent what it now stores.