Add the project path when the compilation database has started loading, instead of after it has finished loading
This makes the behaviour for loading a single compilation database
consistent other situations, including:

* Opening a workspace path, containing a single compilation database
that is still being loaded.
* Opening a workspace path, containing multiple compilation databases,
of which some are already loaded.

This change allows rc calls to be made on files in the project path that
are not yet indexed, but the results from those calls will be suppressed.